add routes and seeders

+++ migrations/08-create-student-payment.js
@@ -48,7 +48,7 @@ module.exports = {
             allowNull: false,
           },
           price: {
-            type: Sequelize.DOUBLE(6, 2),
+            type: Sequelize.DOUBLE,
             allowNull: false,
           },
           compensationAmount: {
