+++ models/teacher_payment.js
@@ -48,7 +48,7 @@ module.exports = (sequelize, DataTypes) => {
         allowNull: false,
       },
       price: {
-        type: DataTypes.DOUBLE(6, 2),
+        type: DataTypes.DOUBLE,
         allowNull: false,
       },
       active: {
